+++ packages/polaris-viz/src/components/LineChart/stories/LineChart.stories.tsx
@@ -113,11 +113,6 @@ Default.args = {
     labelFormatter: formatXAxisLabel,
   },
   yAxisOptions: {labelFormatter: formatYAxisLabel},
-  tooltipOptions: {
-    keyFormatter: (val) => `label: ${val}`,
-    valueFormatter: (val) => `value: ${val}`,
-    titleFormatter: (val) => `title: ${val}`,
-  },
   showLegend: true,
 };